First self-made store by draft
